web_app/requirements.txt
flask
tensorflow
numpy
pillow
werkzeug
PyJWT
pymongo[srv]
dnspython
